Add files via upload
v2

+++ src/main/java/com/example/escaping/buscador/ProvinciaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.example.escaping.data.model.Provincias;
 
-public interface ProvinciaRepository extends JpaRepository<Provincias, Long> {
+public interface ProvinciaRepository extends JpaRepository<Provincias, Integer> {
 //    @Query("SELECT new AlojamientoDTO(h.nombre, h.imagenUrl) FROM Hotel h JOIN h.localidad l WHERE l.provincia.id = :provinciaId")
     // List<AlojamientoDTO> findAlojamientosByProvinciaId(@Param("provinciaId") Long provinciaId);
 }
